Extensions moved to common file, correct build.gradle, remove BaseViewModel,
