--- app/src/main/java/by/mankevich/currencyexchanger/core/presentation/BaseViewModel.kt
@@ -1,5 +1,0 @@
-package by.mankevich.currencyexchanger.core.presentation
-
-import androidx.lifecycle.ViewModel
-
-abstract class BaseViewModel : ViewModel()
