+++ app/src/main/res/layout/currency_exchange_activity.xml
@@ -42,7 +42,6 @@
         app:layout_constraintStart_toStartOf="@+id/guideline2"
         app:layout_constraintTop_toBottomOf="@+id/balance_container" />
 
-
     <ImageView
         android:id="@+id/sell_image_view"
         android:layout_width="48dp"
@@ -68,7 +67,7 @@
 
     <androidx.appcompat.widget.AppCompatSpinner
         android:id="@+id/sell_currency_types_spinner"
-        android:layout_width="100dp"
+        android:layout_width="105dp"
         android:layout_height="0dp"
         app:layout_constraintBottom_toBottomOf="@+id/sell_image_view"
         app:layout_constraintEnd_toEndOf="parent"
@@ -126,7 +125,7 @@
 
     <androidx.appcompat.widget.AppCompatSpinner
         android:id="@+id/receive_currency_types_spinner"
-        android:layout_width="100dp"
+        android:layout_width="105dp"
         android:layout_height="0dp"
         app:layout_constraintBottom_toBottomOf="@+id/receive_text"
         app:layout_constraintEnd_toEndOf="parent"
